+++ saleor/static/dashboard-next/js/components/Table.tsx
